Inspector now renders component name and field names prettier. Added a CamelCaseToWords method in strings that inserts spaces before each new word in a camel case string. Renamed String::FirstCharToUpper -> String::Capitalize

+++ code/addons/physicsfeature/managers/physicsmanager.cc
@@ -68,7 +68,7 @@ PhysicsManager::InitPhysicsActor(Game::World* world, Game::Entity entity, Physic
         Physics::CreateActorInstance(resId, worldTransform, (Physics::ActorType)actor->actorType, Ids::Id64(entity));
     actor->actorId = actorid.id;
 
-    if (actor->actorType == Physics::Kinematic)
+    if ((int)actor->actorType == Physics::Kinematic)
     {
         world->AddComponent<PhysicsFeature::IsKinematic>(entity);
     }
